requirements.txt
requests
gevent
Flask
pymongo
python-dotenv
fuzzywuzzy[speedup]
dnspython 
loguru
